+++ docs/tutorial/example/test_2.cpp
@@ -1,6 +1,6 @@
 ///////////////////////////////////////////////////////////////////////////////
 //
-// file: calc_pi_example_3_gtest.cpp
+// file: test_2.cpp
 // 
 // Simple example that calculates pi via simple integration.
 //
@@ -8,17 +8,40 @@
 // https://www.mcs.anl.gov/research/projects/mpi/usingmpi/examples-usingmpi/simplempi/cpi_c.html
 ///////////////////////////////////////////////////////////////////////////////
 
-#include <iostream>
-#include <cstdlib>
-#include <math.h>
-
 #include <gtest/gtest.h>
 
+#include "mpi.h"
+
 #include "calc_pi.hpp"
+#include "calc_pi_mpi.hpp"
 
-
+// test serial lib
 TEST(calc_pi_tests, serial_example)
 {
     double PI_REF = 3.141592653589793238462643;
     ASSERT_NEAR(calc_pi(1000),PI_REF,1e-1);
+}
+
+
+// test mpi lib
+TEST(calc_pi_tests, mpi_example)
+{
+    double PI_REF = 3.141592653589793238462643;
+    ASSERT_NEAR(calc_pi_mpi(10000),PI_REF,1e-2);
+}
+
+// main driver
+int main(int argc, char * argv[])
+{
+    int result = 0;
+
+    ::testing::InitGoogleTest(&argc, argv);
+
+    MPI_Init(&argc, &argv);
+
+    result = RUN_ALL_TESTS();
+
+    MPI_Finalize();
+
+    return result;
 }
